# packages/contracts/openapi/hospital-api.yaml
openapi: 3.1.0
info:
  title: Hospital Management Platform API
  version: 2.0.0
servers:
  - url: /api/v1
paths:
...
  /doctors/{doctorId}/availability:
    get:
      security:
        - bearerAuth: []
      summary: List doctor availability slots
  /doctors/me/profile:
    get:
      security:
        - bearerAuth: []
      summary: Read doctor profile
    patch:
      security:
        - bearerAuth: []
      summary: Update doctor profile
  /doctors/me/schedule:
    get:
      security:
        - bearerAuth: []
      summary: Read weekly schedule and exceptions
    put:
      security:
        - bearerAuth: []
      summary: Replace weekly schedule
  /doctors/me/exceptions:
    post:
      security:
        - bearerAuth: []
      summary: Create a doctor schedule exception
...
  /admin/doctors:
    get:
      security:
        - bearerAuth: []
      summary: List doctors for admin
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT